[IndexCleaner] ignore non-existent entries inside config
Summary:
If a folder was added to include/excludeFolders and then removed,
there will be a stale entry in the config, for which `filePathToId`
yields 0. When passed to `removeRecursively`, it might reinterpret it
as tree root, which is not something we want (although there are
checks for it inside `WriteTransaction`)

See also {D22502}

Reviewers: #baloo, bruns, ngraham

Reviewed By: #baloo, ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D22557

+++ src/file/indexcleaner.cpp
@@ -75,12 +75,14 @@ void IndexCleaner::run()
     const auto includeFolders = m_config->includeFolders();
     for (const QString& folder : includeFolders) {
         quint64 id = filePathToId(QFile::encodeName(folder));
-        tr.removeRecursively(id, shouldDelete);
+        if (id > 0) {
+            tr.removeRecursively(id, shouldDelete);
+        }
     }
     const auto excludeFolders = m_config->excludeFolders();
     for (const QString& folder : excludeFolders) {
         quint64 id = filePathToId(QFile::encodeName(folder));
-        if (tr.hasDocument(id)) {
+        if (id > 0 && tr.hasDocument(id)) {
             tr.removeRecursively(id, shouldDelete);
         }
     }
